% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saxton_1986.R
\name{saxton_1986_theta_s}
\alias{saxton_1986_theta_s}
\alias{saxton_1986_K}
\alias{saxton_1986_Ksat}
\title{Estimate Hydraulic Conductivity using Saxton et al. (1986)}
\usage{
saxton_1986_theta_s(sand, clay)

saxton_1986_K(sand, silt, clay, theta)

saxton_1986_Ksat(sand, silt, clay)
}
\arguments{
\item{sand}{_numeric_; sand percentage}

\item{clay}{_numeric_; clay percentage}

\item{silt}{_numeric_; silt percentage}

\item{theta}{_numeric_; water content percentage}
}
\value{
_numeric_ estimates of hydraulic conductivity at specified water content given sand, silt and clay percentages
}
\description{
Estimate Hydraulic Conductivity using Saxton et al. (1986)
}
\examples{
# estimate saturated water content 80\% sand, 10\% clay
theta_s <- saxton_1986_theta_s(80, 10)

# estimate hydraulic conductivity at 50\% porosity filled with water
saxton_1986_K(80, 10, 10, theta = theta_s / 2)

# estimate hydraulic conductivity at saturation
saxton_1986_Ksat(80, 10, 10)

}
\references{
Saxton, K.E., W.J. Rawls, J.S. Romberger, R.I. Papendick. Estimating generalized soil-water characteristics from texture. Soil Sci. Soc. Am. J. 50: 1031-1036.
}
