% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saxton_1986.R
\name{saxton_1986_psi}
\alias{saxton_1986_psi}
\title{Estimate Matric Water Potential using Saxton et al. (1986)}
\usage{
saxton_1986_psi(sand, clay, theta)
}
\arguments{
\item{sand}{numeric; sand percentage}

\item{clay}{numeric; clay percentage}

\item{theta}{numeric; water content proportion}
}
\value{
numeric estimates of matric water potential (kPa) at specified water content (`theta`) given `sand` and `clay` percentages
}
\description{
Estimate Matric Water Potential using Saxton et al. (1986)
}
\examples{

theta <- seq(0.00, 1, 0.01)
plot(log10(saxton_1986_psi(80, 10, theta)) ~ theta,
     xlab = "Water Content (\u03b8)", ylab="Log10 Matric Potential (\u03a8, kPa)", 
     type="l", lty=2, xlim=c(0,1))
lines(log10(saxton_1986_psi(35, 40, theta)) ~ theta, lty=3)
legend("topright", legend=c("80\% sand / 10\% clay", "35\% sand / 40\% clay"), lty=2:3)

}
\references{
Saxton, K.E., W.J. Rawls, J.S. Romberger, R.I. Papendick. Estimating generalized soil-water characteristics from texture. Soil Sci. Soc. Am. J. 50: 1031-1036.
}
